feat(wasm-utxo): move fixed script wallet tests to separate file
Move the wallet key address derivation tests from utxolibCompat.ts to a
new fixedScript/address.ts file for better organization and separation
of concerns.
